GlobalWatchlist extension v0.1 - just a shell
Adds the special pages to load the javascript for the front-end,
but no javascript yet (for now just loads a placeholder script
that logs to the console to confirm that it was loaded)

Existing messages were added in i18n/script/en.json so that translation
can begin on those once this is merged, without needing to wait until
the script begins to use them.

This is nowhere near complete, but its a start, and a good break point
to separate the javascript migration from the basic framework.

Bug: T255172
Change-Id: I2ac5314ebcb898593497082c5b7b4b1335d89706