Add settings form on Special:GlobalWatchlistSettings
The form retrieves the user's settings and renders the settings form,
including pre-filling the user's current options.

Retrieving the settings is currently done from
`window.GlobalWatchlistSettings`[1]. It is handled in its own
module, so that it can be used for the code for Special:GlobalWatchlist
as well.

Hitting save results in the various elements being analysed and an
object containing the new preferences to be created. It then delegates
to a separate dedicated module to save the settings.

That saving module it is not yet wired up to make an API call and
actually save the settings, but it logs the new settings to the console
to allow verification that the inputs were interpreted successfully.

[1] Sometimes the script is loaded first and the settings are undefined,
refreshing usually fixes that

Bug: T255221
Change-Id: I4c186d872f5fd227f899287fb4791407b3c80aeb