saveSettings.js - actually call api to update settings
Rewrite the saveSettings module to call the globalwatchlist
api with the user's new settings, and handle the result.

Currently, the result will always be successful, but it will soon
fail if the preferences chosen are invalid (eg no sites chosen).

The settings manager then delegates to a new module for notifications
to tell the user that the saving either succeeded or failed. A followup
will be, once the settings validation is added to the backend, to
interpret the failure and alert the user to the problem with their
choices, rather than just a generic failure message.

Also includes a minor fix for the settings module's handling of the
user's current settings when pre-filling and resetting the form.

Bug: T255221
Change-Id: If6d0b91deffac06570a7bd619c2caddd8dea0352