Special:GlobalWatchlist - create front end for basic global watchlist
The global watchlist code reads a user's preferences (currently from
window.GlobalWatchlistSettings, but that is an implementation detail
and the global watchlist code only interacts with it by calling
GetConfig()) and for each site, renders a basic watchlist.

For testing purposes, it currently shows all changes on a user's
watchlists that match the filters, rather than only unseen changes.
This will be remedied in a subsequent commit once its demonstrated
that the current handling is sound.

Not yet working:
* Autorefresh when all sites are marked as seen
* Wikidata labels being fetched (disabled until it can be tested)
Those will be addressed in follow-up patches.

Testing:
* Basic tests added for linker.js
Follow-ups will include adding tests for at least utils.js

Bug: T255536
Change-Id: I03d79c702c3515c8320c9fb72fe16bbd6185cf45