Grouping edits: combine entries for new pages with edits
If a page is created and then has edits, previously grouping the edits
would have resulted in two entries, one for the new page and one
for the edits to it. Instead, combine these by treating page creations
as edits with an extra flag, rather than a different type of change.

Also remove legacy reference to %SITE% and cleanup outdated handling
from before entries were normalized to include needed fields.

Bug: T257428
Change-Id: I7dcdc553280d7693bc96eaa9f3d8b51ce504d17a