Move wikidata label handling out of site.js
Introduce wikibase.js for getting labels
site.js is still responsible for determining if it should be used

While the new code in wikibase.js is a lot bigger than the code removed from
site.js, the functionality should be identical (documentation was added,
separate methods were factored out for clarity, etc.) - it should be a lot
easier to understand though.

Additionally:
* Fixed bug where wgNamespaceIds cannot be used since they are
fetched in the context of the local wiki. Instead, rely solely on
the title.
* Moved wgGlobalWatchlistWikibaseSite to getSettings

Bug: T258452
Change-Id: I84150261d19c22a784d0eb8d458490145f42fc0d