Add $wgGlobalWatchlistUseVue
When enabled, visiting Special:GlobalWatchlist
should result in a line being printed to the console
that the vue version of the javascript has loaded, and
text to the same effect being shown.

The vue version is not yet implemented, and
when the config flag is enabled Special:GlobalWatchlist
will no longer work. A future patch will actually
implement the Vue version of the special page;
this patch just adds the infrastructure for it.

Bug: T259403
Change-Id: If1747548a64f63883731d56c3d211d1053944954