Fix allEmptySites template usage
The }}} in the comment was interpreted as the end of
the comment

Additionally, at some point since it was added,
the handling of trying to append the jQuery element
broke, and it is instead shown as `[object Object]`.
Use outerHTML to ensure that the HTML of the element
is properly added.

Change-Id: Ic7bb5cd2f3b90f182785a1f4d2d07118204912f1