Special:GlobalWatchlistSettings - 'enter' should not remove site rows
Due to the "implicit submission" feature of html forms, hitting enter
triggers the first submit button in the form, which in this case is the
button associated with the top site row, resulting in that button's
intended functionality (remove the row) being executed
unexpectedly. Solve this by adding another submit button before
the site rows, so that that button is triggered instead. Disable the
button so that its triggering does not result in the form being
submitted, and then hide the button.

To hide the button even when the viewer does not have javascript
enabled, make the relevant ResourceLoader module,
ext.globalwatchlist.specialglobalwatchlistsettings, loadable
via OutputPage::addModuleStyles() by removing its dependency
on `mediawiki.htmlform.ooui` - instead, manually load that
module separately.

Loading the styles separately from the `mediawiki.htmlform.ooui`
module causes our styles to be loaded first, and *then* the
core styles which we are trying to override. Make sure our
styles take precedence by making the selectors more specific,
converting the file to LESS to make that easier.

I'm not sure why, but for some reason this fix only works
for hitting enter from within the text inputs - hitting
enter while selecting one of the radio or checkbox options
still triggers the removal of the first site row. I could
not fix those, trying with

* Combining the different sections of the form into one
* Adding a dummy submit button to each section
* Messing with the attributes of the dummy submit button itself

Maybe eventually we can fix what happens if you hit enter
elsewhere, but this patch is still an improvement, because

* in the current version, hitting enter for a radio or checkbox
option triggers the removal, so this hasn't changed
* enter is most likely to be hit when typing in the site rows,
which this patch does fix

Bug: T275588
Change-Id: I5ad9a351b1b9c989d1d4e3d8c299bbfd3d29beaf