Selenium tests for normal version of Special:GlobalWatchlist
At the start of the tests, log in to the existing admin account,
for the bot, make an edit to a page, making sure to watch the
page in the process, and then set the notification timestamp
to earlier (we set it to 1 June 2021 to guarantee its from
before the edit was made, but any time before the edit would
work). We need to set the timestamp back because the global
watchlist only shows unseen changes (unlike Special:Watchlist
which includes seen changes too by default). This part will
be moved into a `before` hook in a follow-up, to be reused
for the eventual vue tests.

We then open Special:GlobalWatchlist load the normal display
mode, and check that the relevant button type is correctly shown.
Wait for the display to finish loading the data[1] and then check
for the correct entry at the top of the page list.

By default, the only site shown in Special:GlobalWatchlist
is the local site - this is good, its unlikely we will be able
to test the crosswiki functionality in selenium. It also means
that we know that the first site displayed will be the correct
wiki.

We'll add the tests for the vue version of the display in a
separate commit, to make things easier.

[1] We can't use whether the loading indicator is being shown
because it isn't shown immediately, so instead check for
the display of the first site entry

Bug: T284521
Change-Id: I7fed396ec73fb6264696fb09ac2e805796efc6e6