Special:GlobalWatchlist selenium tests - move setup to `before`
Prevent duplication between the tests for the normal and vue
versions of the display by moving the setup processing to
a `before` handler. This covers
- Logging in to the display
- Getting the random title string
- Logging in to the bot, making the edit, and resetting
the notification timestamp

This needs to be done *before* adding the separate vue
tests so that we don't create two edits, because the
order of them being displayed might change.

Bug: T284521
Change-Id: Id3e6a8b8d824a7978298380cd2303faac2ca17e9