Selenium tests for vue version of Special:GlobalWatchlist
To match the existing tests for the normal version.

Requires setting $wgGlobalWatchlistDevMode to true to
make sure that the displaymode parameter is obeyed.
Doing this in the extension.json default until there
is a better way (T248304/T267928). This shouldn't
affect uses in WMF production, because we explicitly
set it to false.

By checking for a button with the `wvui-button` class, this
likely would have caught the regression in T284198#7130949
earlier.

Bug: T284521
Change-Id: I7c238a21725ea09d73a65faca16863f25b12e99f