Make watchlistUtils non-static
Rename to GlobalWatchlistWatchlistUtils and move the file to
modules/WatchlistUtils.js, following existing naming scheme
where the class is `GlobalWatchlist{name}` and the file name
doesn't include the "GlobalWatchlist".

The linker is now injected instead of being provided as
a parameter to rawToSummary() and then passed around.

Additionally, remove the `site` parameter, unused.

Required adding GlobalWatchlistWatchlistUtils to the
always defined types for eslint rule jsdoc/no-undefined-types;
because we don't directly import (via require()) the class
everywhere that we use it, jsdoc doesn't realize that the class
is defined in another file.

Bug: T286141
Change-Id: Ie268e19c7f767a307624ffab7a95d3047af82c46