Remove MultiSiteWrapper.markAllSitesSeen() confirmation parameter
The MultiSiteWrapper is created with a configuration object that
already includes the relevant preference, and it does not change
over the course of a visit to Special:GlobalWatchlist. Retrieve
and save the preference in the constructor instead of needing
to pass it.

Also add documentation about why we *cannot* do the same with
the configuration passed to MultiSiteWrapper.getAllWatchlists()

Bug: T286268
Change-Id: I7a5e508fdeb621f6558590c2e9233d93ca8a159b