Add initial storybook documentation
Starting off with documentation for the Toggle and
CollapsibleWrapper components. I used the
storybook documentation for wvui as a reference point
for the versions of the storybook packages and addons to use.

We need to have wvui as a dev dependency so that storybook
and webpack understand the components that we retrieve from
ResourceLoader that come from wvui.

Its fairly basic and just for 2 components, but once this is set up
I'll send more patches for the rest.

Requires updating phpcs because otherwise it tries to run php
sniffs on the added javascript code for some reason. Tell it
to only care about files ending in "php".

Bug: T286272
Change-Id: I6a3e4337cb4623a77761433ce2a5ecc001479114