Remove debug logging from SiteDisplay.makePageLink()
At the end of the function, a debug entry is created with two
elements, the input `entry` and the created `$row`. The intention
behind this was to make it easier to debug a weird looking row by
checking that the jQuery object to display it was created properly.

Checking how it appears in the actual log in my browser, however,
shows that the jQuery object isn't included in the log. Thus, all
we are really getting is the data that was being used to create
the row, logged once for each and every row that is display in
the jQuery/OOUI version of the display. However, we already
have a log entry of all of this data together:
* SiteDisplay.makePageLink() is called in
SiteDisplay.renderWatchlist() for each of the entries in the
array that is passed
* SiteBase.getWatchlist() calls .renderWatchlist(), which is
overridden in SiteDisplay, right after creating a debug
entry with that overall array, labelled "getWatchlist summary"
accordingly, the only benefit that comes from the existing
log entries in makePageLink() is that we get a separate entry
for the data for each row, but given that this debug log would
only be consulted in case of errors, the effort needing to locate
the relevant entry in the overall array is not so high as to warrant
the additional individual log entries.

Another reason to keep the overall log entry instead of the individual
ones is because that overall log entry is also included in the vue
display.

Bug: T286306
Change-Id: Iea0a5ba17330a9156a664f410cfea25df7b02add