getAssociatedPageTitle: use action=query&prop=info instead of parser
When a page is unwatched, a visual indicated is added (the entry
row is struck through) and any entry for the associated talk page
should also be updated (or the associated subject page if a talk
page is unwatched). We currently fetch the name of that associated
page via the parse api, which has some issues regarding escaping.

Instead, switch to use the new "associatedpage" property in the
info query, which should not have these escaping issues and
is also intended to do this, rather than using the parser with
magic words {{TALKPAGENAME}} and {{SUBJECTPAGENAME}}.

Requires bumping the required version of MediaWiki to 1.37+
because the "associatedpage" property was just added (see
T257014).

Bug: T278713
Change-Id: I6484e7cee2c4f384de4b21226af52afd278509d2