Remove debug logging from vue markSiteAsSeen method
In SpecialGlobalWatchlist.vue when the button to mark a
specific site as seen is clicked, we send a debug entry about
that from the vue code, before then calling SiteBase.markAsSeen()
which already sends its own debug entry that is clearer since it
follows the normal format of including the site before the
message, and is also used in the display version. To eliminate
the duplicate messages, remove the extra one in the vue display.

Bug: T286306
Change-Id: Id9459764ede5b270174f794a1616f004fecee3f3