Move tags display management to WatchlistUtils
We already have a shared method to create the display
for a given entry (SiteBase.getTagsDisplay) but that
required both versions of the display to manually call it,
especially tricky for the Vue display where it has to be called
in SiteVue because its not available in the rendering.

Instead, add the logic to WatchlistUtils so that the entries
that creates already have the display html too.

Bug: T286268
Change-Id: I3f94518b4d95ca33636045baf1a560d13e60059c