Display: Use HTML "dir" attribute for ltr/rtl
mw-content-ltr and -rtl classes are no longer enough to
do this automatically in the new vector, since that is a part
of the "legacy" styles.

Not updating the Vue version, that will be handled separately
in T274603 and was not working earlier anyway.

Bug: T287649
Change-Id: I66107dca7ab3d072ee1a712eaa5228bb318e0ce5
(cherry picked from commit 03f43fcda98fb669566450d244fceef535163999)