Deduplicate timestamp title handling
In WatchlistUtils set the timestamp title to null by default,
and then update it to reflect grouped changes when needed.
Both jQuery and Vue ignore null attributes.

Also a bit of cleanup to the Vue EntryRow component, no need
to defined computed properties for the timestamp when the
entry can be accessed directly.

Bug: T286268
Change-Id: Ia794cf7043a984c46ce8bc1856ab8be302491edc