WatchlistUtils: merge rawToSummary end cleanup
In preparation for adding new classes to represent entry
rows, either log entries or edits, we need to keep the
two groups separate when they have their final processing
in rawToSummary, so that we can eventually replace the
arrays of plain objects with arrays of the new classes to be
introduced. To avoid needing to call each of the current
individual cleanup methods for both groups, merge those
methods.

- addExpirationMessages
- addEntryFlags
- truncateTimestamps
- addTagDisplays
- addCommentDisplays

Refactor the tests to allow checking just a part of the new
getFinalEntries() method at a time, by adding a helper function
to assert that an object has a subset of properties that match
the expected result.

This should have no net impact on the results of rawToSummary()

Bug: T288385
Change-Id: Ib3965698dcb6bbbeba66bf5c0a494db529b61f0d