Drop abandonware Vue version, unmaintained for over a year
WVUI is not to be used. We've waited over a year for the code to be
migrated, but there's no activity and this code isn't enabled anywhere,
so dropping it for now. When development time is available to migrate
from WVUI to Codex, this patch can be reverted as the starting point.

Bug: T310240
Change-Id: Ib0bff44fec61730b7480f323fb4d2d7d58edbc09