Move GWGuidedTourHooksTest::testNewFromGlobalState to an integration test
"Global state" and "UnitTest" can no longer appear in the same class; in
this particular case, the call to ExtensionRegistry::getInstance()
throws an exception.

Create a new integration test class for this method. It doesn't really
do much for now, but someone may want to expand it at some point.

Change-Id: I38e246a3b18afb563b1c35104a27582be78b367f