Replace hardcoded wiki directions list with live data
Why:

Previously, the directionality of wikis (left to right or right to left)
used to be determined by a regex matching the site ID. This was fragile
on WMF wikis (new RTL languages would not be recognized) and completely
unusable on wiki farms not using the same domain name structure.

What:

Replace the regex with code that automatically queries this data from
the wiki. Since it’s bundled with the first watchlist query, it means no
extra HTTP requests and thus should have minimal impact on the
performance. (Maybe the list of change tags could also be queried in the
same request to even reduce the number of requests. However, that’s for
a future patch.)

Bug: T412505
Change-Id: I09eed1adfb7056fc0773e1db067c070d35bc8424